Fixed: ToDoListItem has a Circle Avatar with abbreviation
For some reason ToDoListItem has text test passed with no problems; however, The circle Avatar returned the incorrect color of black. color.black instead of color.black54. The abbreviation and the actual name of the item in the list were switched around. Now, the abbreviation is in the avatar while the full name is in the list.