Clicking and Typing adds item to ToDoList Test Passed
First, the "OKbutton" was misinput as "Okbutton". Next, Rather than be connected to when OKbutton was pressed, the method to _handleNewItem was connected to the Cancel Button instead. 
Finally, in _handleNewItem, the Item item was set to equal a constant value of Item with its name value being "ItemText" rather than the actual variable ItemText.