Re: Fixed Cancel Button + Test for Editing an Item Added
Fixed Cancel Button by removing if statement and replacing it with syntax similar to OKbutton. 
Test for Editing an Item is now implemented. Test thoroughly goes through same steps a user would to edit an item in the list. 
By comparing to previous tests, this test should work correctly and as such, has passed.